+++ src/app/app-routing.module.ts
@@ -4,7 +4,7 @@ import { AdminComponent } from './components/admin/admin.component';
 import { CarRegisterComponent } from './components/car-register/car-register.component';
 import { DriverRegisterComponent } from './components/driver-register/driver-register.component';
 import { LoginComponent } from './components/login/login.component';
-import { DriverComponent } from './driver/driver.component';
+import { DriverComponent } from './components/driver/driver.component';
 
 
 const routes: Routes = [{
