--- src/app/cars.ts
@@ -1,4 +1,0 @@
-export class Cars{
-    carId: number;
-    seats: number;
-      }
