+++ src/app/services/mark-inactive-driver.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MarkInactiveDriverService } from './mark-inactive-driver-service/mark-inactive-driver.service';
+
+describe('MarkInactiveDriverService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: MarkInactiveDriverService = TestBed.get(MarkInactiveDriverService);
+    expect(service).toBeTruthy();
+  });
+});
