--- src/app/table-filter-pipe.ts
@@ -1,14 +1,0 @@
-import { Pipe, PipeTransform } from '@angular/core';
-
-@Pipe({
-  name: 'tableFilter'
-})
-export class TableFilterPipe implements PipeTransform {
-
-  transform(list: any[], value: string) {
-  
-
-    return value ? list.filter(driver => driver.driver === false) : list;
-  }
-
-}
